fix: chrome 134+ does not reload extensions without developer mode enabled
chrome://extensions developer mode is required at all times for extensons to work. It works with --load-extension but resets after the extension is reloaded chrome.runtime.reload() which is exactly the behaviour that crxjs depends on and is required for e2e tests to work. '--disable-features=ExtensionDisableUnsupportedDeveloper' fixes the issue. https://github.com/microsoft/playwright/issues/34711#issuecomment-2797778673