fix(tests): use separate Vite caches for build and serve tests
Fixes issues where they could interfere with each other when running at
the same time. This also means Playwright gets separate Chromium profile
directories for the different tests.

Note that changing the cachedir for dev servers actually changes the
resulting layout of the built files for some reason, so that is left
the same.